Homework 1: Solution 1
This solution for Homework 1 involves trying to use cin.get(char* s, streamsize) to get the first letter of the answer while ignoring everything else after that, so that I can use 1 cin.ignore before asking for their grade.

As of now, I have to use as many cin.ignores as there are extra charaters. So if the answer is 'yes' , I have to use 2 ignores to ignore 'e' and 's'. But if the answer was 'no', I would only need 1 ignore.